+++ src/primitives/sphere.js
@@ -11,7 +11,7 @@ import { getPopulatedCoordinateBuffer } from '../bufferUtils.js';
 
 // See https://reference.wolfram.com/language/ref/Sphere
 // for the high-level description of what is being rendered.
-export default function ({ color, coords, opacity = 1, radius = 1 }, extent) {
+export default function ({ color = [1, 1, 1], coords, opacity = 1, radius = 1 }, extent) {
 	const sphereGeometry = new InstancedBufferGeometry().copy(
 		new SphereGeometry(radius, 48, 48)
 	);